requirement.txt
asttokens==3.0.0
blinker==1.9.0
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
comm==0.2.2
contourpy==1.3.2
cryptography==44.0.0
cycler==0.12.1
dash==2.17.1
dash-bootstrap-components==1.5.0
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-iconify==0.1.2
dash-table==5.0.0
dash_mantine_components==2.1.0
debugpy==1.8.12
decorator==5.1.1
et_xmlfile==2.0.0
executing==2.2.0
Flask==3.0.3
flask-cors==4.0.0
fonttools==4.58.5
greenlet==3.1.1
idna==3.10
importlib_metadata==8.7.0
ipykernel==6.29.5
ipython==8.32.0
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.8
MarkupSafe==3.0.2
matplotlib==3.10.3
matplotlib-inline==0.1.7
mysql-connector-python==9.2.0
nest-asyncio==1.6.0
numpy==2.3.1
openpyxl==3.1.5
oracledb==2.5.1
packaging==24.2
pandas==2.3.1
parso==0.8.4
pexpect==4.9.0
pillow==11.3.0
platformdirs==4.3.6
plotly==5.17.0
prompt_toolkit==3.0.50
psutil==7.0.0
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pure_eval==0.2.3
pycparser==2.22
Pygments==2.19.1
PyMySQL==1.1.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
PyJWT==2.8.0
pyzmq==26.2.1
requests==2.32.4
retrying==1.4.0
seaborn==0.13.2
setuptools==80.9.0
six==1.17.0
SQLAlchemy==2.0.37
stack-data==0.6.3
tenacity==9.1.2
timeout-decorator==0.5.0
tornado==6.4.2
traceback-with-variables==2.2.0
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.5.0
wcwidth==0.2.13
Werkzeug==3.0.6
xlsxwriter==3.2.5
zipp==3.23.0
